backend/requirement.txt
fastapi>=0.95.0
uvicorn[standard]>=0.22.0
pydantic>=1.10.0
python-dotenv>=1.0.0
psycopg2-binary>=2.9.6
python-multipart>=0.0.6
SQLAlchemy>=1.4.0
passlib[argon2]
python-jose[cryptography]>=3.3.0
pydantic[email]
httpx>=0.27.0
kafka-python-ng

# --- AI / Machine Learning ---
transformers
torch
scipy
datasets
numpy